Add an installer Dockerfile
This Dockerfile describes a multi-stage build that downloads a Racket
prerelease Linux installer script and runs it in “in place” mode,
resulting in a Docker image with `racket`, `raco`, `scribble`, etc.
available from the command line. Build arguments allow customizing the
base image used as well as the specific installer variant downloaded.
Unix-style installs are not included yet due to difficulty in ensuring
the installer itself is not included in the final image’s layers.