Add a docker-compose configuration for running tests
This tests four variants of a prerelease installer: one for each
combination of a full or minimal distribution for either the x86_64 or
i386 platforms. For each variant, the test verifies that:

- An in-place install completes successfully
- Racket code can be evaluated
- Packages can be installed over HTTPS
- The working non-GUI tests pass

Tests still need to be included for the following:

- Unix-style installations
- Generated uninstall scripts
- From-source builds, both with and without pre-built packages
- GUI tests
- `linux-natipkg` installers (which seem to have OpenSSL issues)

Unfortunately, `docker-compose up —build` doesn’t work as nicely as it
should because that runs tests in parallel, resulting in four parallel
runs of the intensely memory-hungry database tests. Results in
termination of containers with extreme prejudice.