Add GUI test services
The GUI test services all share a single xvfb service with only one
screen. As a result, GUI tests cannot be run in parallel. The easiest
way to fix that would be to run xvfb with multiple screens and
configure each GUI service to use a specific screen. The robust fix
would be to change the endpoint_mode of xvfb from a virtual IP to a DNS
Round Robin, then hook that up with some sort of “X server load
balancer” that dynamically discovers xvfb containers and allows clients
to point themselves at the load balancer with DISPLAY and automatically
get their own isolated X display. That’s more magic than I’m going to
bother with for now.