Add a way to watch the running GUI tests in a browser with noVNC
This uses noVNC, a web app for viewing a VNC server in a browser, to
watch what the GUI tests are doing to the xvfb service as they run.
This works via a chain of proxies:

- As defined in the previous commit, the x11vnc service proxies the
xvfb service into a VNC server.
- The websockify service proxies the VNC server to tunnel VNC traffic
over the Websocket protocol, making it available to browsers.
- The nginx service provides both static HTML for the noVNC client app
and a special resource named `/websockify` that noVNC assumes is a
Websocket-accessible VNC server.

The noVNC client app is downloaded from GitHub during image building,
but mounted into nginx by way of a volume to avoid another dockerfile
for building a derived nginx. Additionally, the official nginx docker
image is intended to be used with hot-reloaded static files and
configuration by way of the special `/usr/share/nginx/html/` and
`/etc/nginx/conf.d/` directories. Cooperating with this intention makes
configuration easier and means nginx doesn’t has to be restarted. Using
docker configs instead of volumes might be better (avoids complexity
with volume persistence assumptions) but this works for now.