Combine together GUI and non-GUI test dockerfiles
They’re still run as separate services in docker-compose, but now the
same Dockerfile and image is used (just with different commands). This
lets use use the USER instruction to set the user to `racket` instead
of `root` without messing up installation of GUI packages with
`apt-get`. Also adds a YAML anchor for the GUI test service to make the
configuration a little less repetitive.