Wrap test building and running inside a "test runner" service
Instead of defining, building, and running each set of test services
for each installer we define a single “test runner” service. This test
runner contains a parameterized docker compose configuration for
building and running the tests of a single installer variant. The outer
docker-compose specifies that the containerized test runner’s
docker-compose should mount the host docker socket, ensuring that the
test runner uses the host docker engine instead of an isolated docker
engine inside the container. With a small script taking an argument for
which installer to test, we can now build and run all the tests for an
existing installer image (e.g. `racket-x86`) using `docker-compose run
racket-test-runner racket-x86`.