Wrap in-place installation inside an "installer" service
Similar to the test-runner service. The installer service consumes a
Racket installer and executes it, producing an image containing a
Racket installation. This is defined in terms of “installer types” and
“installation types”. An installer type identifies a particular
artifact available for download from a version-specific Racket
installers website such as a 64-bit linux installation script. An
“installation type” is a particular way of executing an installer,
since some installers can be installed in multiple different ways.
Currently the only implemented installation type is “script-inplace”,
meaning an in-place install of a Linux shell script installer. By
wrapping installation in a service, new installation types can be
defined with a linear amount of configuration in the number of
installation types.